Reorganize numeric interface
